* A test is moved to BasicChatStoryTest.java * AnanichTest.java is deleted
